reorganize SysrootConfig and feature handling
Turns out that the sysroot 'root' crate forwards features to std, so we can just
say that SysrootConfig determines the root crate and features apply to that
crate and the rest must be handled on the sysroot side.

In particular we no longer set any default features for `alloc`. The old default
corresponds to adding the `compiler_builtins/mem` feature.